requirements.txt
google-cloud-vision
openai
Pillow
tqdm
requests
beautifulsoup4
PyMuPDF
python-dotenv
Flask
pdf2image
gunicorn
flask-cors 
